refactor: group connection profile helpers
